switch to js to use import.meta.resolve
